+++ src/components/common/AlbumPoster.tsx
