+++ src/components/common/Battles.tsx
