+++ src/components/common/Genres.tsx
@@ -5,7 +5,7 @@ import { getGenres } from '@/components/post/api';
 import { COLOR } from '@/constants/color';
 import useGenre from '@/hooks/useGenre';
 
-import GenreSkeleton from '../skeleton/GenreSkeleton';
+import GenreSkeleton from './skeleton/GenreSkeleton';
 
 interface Props {
   shouldNeedAll?: boolean;
