+++ src/components/common/NoContent.tsx
